Refactor and optimise D3D11 refcounting implementation
* In heavy D3D11 workloads the refcounting overhead especially during fast
  binding changes was significant. Refactoring the refcounting to work on a
  different model and deferring destruction of objects removes most of the
  overhead.

+++ util/test/demos/d3d12/d3d12_large_buffer.cpp
@@ -55,7 +55,7 @@ RD_TEST(D3D12_Large_Buffer, D3D12GraphicsTest)
     verts[indices[1]] = DefaultTri[1];
     verts[indices[2]] = DefaultTri[2];
 
-    range.End = vb->GetDesc().Width;
+    range.End = (SIZE_T)vb->GetDesc().Width;
 
     vb->Unmap(0, &range);
